Changes to SimpleKMeans to improve runtime and a couple of small fixes to SimpleKMeans and EM to fix behaviour for weighted data.
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@11290 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/src/main/java/weka/clusterers/EM.java
@@ -849,8 +849,8 @@ public class EM extends RandomizableDensityBasedClusterer implements
 
     Instances centers = bestK.getClusterCentroids();
     Instances stdD = bestK.getClusterStandardDevs();
-    int[][][] nominalCounts = bestK.getClusterNominalCounts();
-    int[] clusterSizes = bestK.getClusterSizes();
+    double[][][] nominalCounts = bestK.getClusterNominalCounts();
+    double[] clusterSizes = bestK.getClusterSizes();
 
     for (i = 0; i < m_num_clusters; i++) {
       Instance center = centers.instance(i);
